What is Zeus Network？
Zeus Network: A new high-level architecture that aims to bring this future metaverse closer to reality by providing a framework for creating Binance Chain compatible scaling solutions.
Zeus Network proposes a new blockchain architecture based on Proof of History (PoH) - proof of verification order and passage of time.

PoH is used to encode trustless passage of time into a ledger - an append-only data structure. when with
Consensus algorithms such as Proof of Work (PoW) or Proof of Stake
(PoS), PoH can reduce the message passing overhead in Byzantine fault-tolerant replicated state machines, resulting in sub-second finalization times.

At the same time, Zeus Network also proposes an algorithm that utilizes the time-keeping properties of the PoH ledger. It behaves as a PoS algorithm that can be recovered.

From partitions of any size and efficient streaming proof-of-replication (PoRep). The combination of PoRep and PoH provides defense against ledger falsification in terms of time (ordering) and storage. The protocol was profiled on a 1 gbps network, showing that throughputs of up to 900k transactions per second are possible.